requirements.txt
scikit-learn
pandas
joblib
pytest
dvc
dvc-gs
mlflow
fastapi
pydantic
uvicorn
opentelemetry-exporter-gcp-trace
